% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_transformations.R
\name{asinh.transform}
\alias{asinh.transform}
\title{Asinh transformation for InteractionSet}
\usage{
\method{asinh}{transform}(IntSet)
}
\arguments{
\item{IntSet}{An InteractionSet object including two replicates of
Hi-C contact counts.}
}
\value{
A variance-stabilized InteractionSet object.
}
\description{
This function applies asinh transformation
to Hi-C contact counts from two replicates (provided as an
InteractionSet object)
}
