% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_hic.R
\name{cool2IntSet}
\alias{cool2IntSet}
\title{Read .cool/.mcool sparse matrix into a InteractionSet object}
\usage{
cool2IntSet(files, chr1, chr2, res = NULL)
}
\arguments{
\item{files}{The list of paths to a HDF5 stored cools (uni-dimension)
or mcools (multi-dimension sparse matrix). If using a multi-dimesion,
the resolution of one of the dimension need to be passed to res.}

\item{chr1}{The chromosome name including genomic bins of the first anchor.}

\item{chr2}{The chromosome name including genomic bins of the second anchor.}

\item{res}{'NULL' if using a uni-dimensional cool file or the resolution of one of layer in the mcool file}
}
\value{
An InteractionSet of relevant bin pairs with the number of interaction for each pair
}
\description{
This function reads HiC contact matrix file(s) created by the
cooler application (https://github.com/mirnylab/cooler) stored in a
HDF5 data storage and converts them to an InteractionSet object for genomic bins
within a pair of chromosomes with the counts of reads in these bin pairs.
}
