% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_transformations.R
\name{hft.transform}
\alias{hft.transform}
\title{Haar-Fisz variance-stabilizing transformation for InteractionSet}
\usage{
hft.transform(IntSet, chromoR = FALSE)
}
\arguments{
\item{IntSet}{An InteractionSet object including two replicates of
Hi-C contact counts.}

\item{chromoR}{If TRUE, it does denoising of coefficients as in chromoR.}
}
\value{
A variance-stabilized InteractionSet object.
}
\description{
This function applies Haar-Fisz variance-stabilizing transformation
to Hi-C contact counts from two replicates (provided as an
InteractionSet object)
}
