% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAD_functions.R
\name{iset_hicseg}
\alias{iset_hicseg}
\title{Apply HiCseg on InteractionSet}
\usage{
iset_hicseg(iset, rep_num, max_cp = 40, model = "D", dist = "G")
}
\arguments{
\item{iset}{Input InteractionSet.}

\item{rep_num}{The replicate's number in InteractionSet to
be used for TAD annotation.}

\item{max_cp}{HiCseg's parameter: maximum number of
change points.}

\item{model}{HiCSeg's parameter: "D" for block-diagonal
and "Dplus" for the extended block-diagonal model.}

\item{dist}{HiCseg's parameter: Distribution of the data:
"B" is for Negative Binomial distribution,
"P" is for the Poisson distribution and
"G" is for the Gaussian distribution.}
}
\value{
A GRanges object including identified TADs' ranges.
}
\description{
This is a wrapper function for HiCseg to take InteractionSet
object as input and return TADs as GRanges object.
}
