% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAD_functions.R
\name{iset_spectralTAD}
\alias{iset_spectralTAD}
\title{Apply SpectralTAD on InteractionSet}
\usage{
iset_spectralTAD(iset, rep_num, res, window_size)
}
\arguments{
\item{iset}{Input InteractionSet.}

\item{rep_num}{The replicate's number in InteractionSet to
be used for TAD annotation.}

\item{res}{resolution of iset.}

\item{window_size}{SpectralTAD's parameter: The size of
the sliding window for calculating TADs.}
}
\value{
A GRanges object including identified TADs' ranges.
}
\description{
This is a wrapper function for SpectralTAD to take
InteractionSet object as input and return TADs as
GRanges object.
}
