% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAD_functions.R
\name{iset_topdom}
\alias{iset_topdom}
\title{Apply TopDom on InteractionSet}
\usage{
iset_topdom(iset, rep_num, res, window_size)
}
\arguments{
\item{iset}{Input InteractionSet.}

\item{rep_num}{The replicate's number in InteractionSet to
be used for TAD annotation.}

\item{res}{resolution of iset.}

\item{window_size}{TopDom's parameter: The number of bins
to extend (as a non-negative integer).}
}
\value{
A GRanges object including identified TADs' ranges.
}
\description{
This is a wrapper function for TopDom to take InteractionSet
object as input and return TADs as GRanges object.
}
