% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_hic.R
\name{store_IntSet}
\alias{store_IntSet}
\title{Storing IntSet assays}
\usage{
store_IntSet(IntSet, rep_num, filepath)
}
\arguments{
\item{IntSet}{The InteractionSet object to store.}

\item{rep_num}{Number of the assay to store.}

\item{filepath}{The path to write a tab-sep file.}
}
\description{
This function writes interaction frequencies of assay from IntSet into 7-columns
tab-sep format (chr1, start1, end1, chr2, start2, end2, count)
}
