% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_functions.R
\name{VI}
\alias{VI}
\title{The variance instability measurement}
\usage{
VI(IntSet, bin)
}
\arguments{
\item{IntSet}{An InteractionSet object including two replicates of
Hi-C contact counts.}
}
\value{
VI value.
}
\description{
This function calculates a variance instability (VI) measurement based on
observed differences between replicates.
}
