% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vss.train.R
\name{vss.train}
\alias{vss.train}
\title{Learning the empirical mean-variance relationship from
two replicates.}
\usage{
vss.train(
  cts,
  aux_type = "rep1",
  group_type = "bin",
  zero_bin = TRUE,
  bin_size = 100,
  log_scale = TRUE
)
}
\arguments{
\item{cts}{A two-column data.frame including counts from two replicates.}

\item{aux_type}{The type of auxiliary signals in the VSS algorithm.
'rep1' (or 'rep2') uses 'rep1' (or 'rep2') to construct base signals and
'rep2' (or 'rep1') to construct auxiliary signals. 'mix' uses
both replicates to construct each of the base and auxiliary
signals. 'rep1' or 'rep2' should be used when library sizes differ
and data is not normalized.}

\item{group_type}{The strategy to group auxiliary counts to
calculate empirical means and variances. 'bin' divides a sorted
auxiliary counts into equal size bins and calculate mean and
variance in each bin. 'value' groups auxiliary counts based
on their corresponding base counts and calculate mean and
variance in each group. For rare signals (large signal values)
where the size of groups decreases, bin strategy is used for
more robust estimations.}

\item{zero_bin}{Due to excess of zero counts, there are
significant number of bins corresponding to base count = 0.
Including large amounts of such mean and variance estimations
might cause a problem in fitting the mean-variance relationship.
zero_bin = TRUE creates one auxiliary bin corresponding to
zero base counts.}

\item{bin_size}{The size of the bin in a grouping step.}

\item{log_scale}{If TRUE, spline is learned using log-scaled
means and variances.}
}
\value{
A list of two objects, a three-column data.frame
including estimated means, variances, and predicted variances by smooth.spline
and a fitted smooth.spline object.
}
\description{
This function takes measurements from two replicates (as a two-column
data.frame) to calculate observed variances across different signal intensities
by VSS algorithm
(https://academic.oup.com/bioinformatics/article/37/23/4383/6308936).
}
