% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vss.transform.R
\name{vss.transform}
\alias{vss.transform}
\title{vss transformation}
\usage{
vss.transform(cts, vss_fit, log_scale = TRUE, log_converge = FALSE)
}
\arguments{
\item{cts}{A two-column data.frame including counts from two replicates.}

\item{vss_fit}{A smooth.spline object.}

\item{log_scale}{Whether spline is learned using log-scaled
means and variances.}

\item{log_converge}{If TRUE, variance-stabilizing transformation
converges to the log function asymptotically.}
}
\value{
A variance-stabilized InteractionSet object.
}
\description{
This function takes measurements from two replicates and a trained
mean-variance relationship and outputs a two-column data.frame
object including variance-stabilized counts.
}
