% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vssHiC.R
\name{vssHiC.train}
\alias{vssHiC.train}
\title{Learning the empirical mean-variance relationship from
two Hi-C replicates.}
\usage{
vssHiC.train(
  IntSet,
  aux_type = "rep1",
  group_type = "bin",
  zero_bin = TRUE,
  bin_size = 100,
  log_scale = TRUE
)
}
\arguments{
\item{IntSet}{An InteractionSet object including two replicates of Hi-C contact counts.}

\item{aux_type}{The type of auxiliary counts in VSS algorithm.
'One_rep' uses one replicate to construct base signals and
another one to construct auxiliary signals. 'mix' uses
both replicates to construct each of base and auxiliary
signals. 'One_rep' should be used when library sizes are different
and data is not normalized.}

\item{group_type}{The strategy to group auxiliary counts to
calculate empirical means and variances. 'bin' divides sorted
auxiliary counts into equal size bins and calculate mean and
variance in each bin. 'value' groups auxiliary counts based
on their corresponding base counts and calculate mean and
variance in each group. For rare signals (large signal values)
where the size of groups decrease, bin strategy is used for
more robust estimations.}

\item{zero_bin}{Due to excess of zero counts, there are
significant number of bins corresponding to base count = 0.
Including large amount of such mean and variance estimations
might cause a problem for fitting the mean-variance relationship.
zero_bin = TRUE creates one auxiliary bin corresponding to
zero base counts.}

\item{bin_size}{The size of the bin in a grouping step.}

\item{log_scale}{If TRUE, spline is learned using log-scaled
means and variances.}
}
\value{
A list of two objects, a two-columns dataframe
including means and variances used for training and a
smooth.spline object.
}
\description{
This function uses Hi-C contact counts from two replicates
(provided as an InteractionSet object) to calculate
observed variances among different count intensities by VSS
algorithm (https://academic.oup.com/bioinformatics/article/37/23/4383/6308936).
}
