% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vssHiC.R
\name{vssHiC.transform}
\alias{vssHiC.transform}
\title{vss transformation for Hi-C data}
\usage{
vssHiC.transform(IntSet, vss_fit, log_scale = TRUE, log_converge = FALSE)
}
\arguments{
\item{IntSet}{An InteractionSet object including two replicates of
Hi-C contact counts.}

\item{vss_fit}{A smooth.spline object.}

\item{log_scale}{Whether spline is learned using log-scaled
means and variances.}

\item{log_converge}{If TRUE, variance-stabilizing transformation
converges to the log function asymptotically.}
}
\value{
A variance-stabilized InteractionSet object.
}
\description{
This function uses Hi-C contact counts from two replicates
(provided as an InteractionSet object) and a trained
mean-variance relationship and outputs an InteractionSet
object including variance-stabilized counts.
}
