% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_transformations.R
\name{vst.train}
\alias{vst.train}
\title{Learning mean-dispersion relationship from
two Hi-C replicates.}
\usage{
vst.train(IntSet)
}
\arguments{
\item{IntSet}{An InteractionSet object including two replicates of
Hi-C contact counts.}
}
\value{
A dispersionFunction object.
}
\description{
This function learns mean-dispersion function from DESeq2
given Hi-C contact counts from two replicates (provided as an
InteractionSet object)
}
