% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_transformations.R
\name{vst.transform}
\alias{vst.transform}
\title{A variance-stabilizing transformation for InteractionSet based on 'vst'
function from 'DESeq2'}
\usage{
vst.transform(IntSet, dispF)
}
\arguments{
\item{IntSet}{An InteractionSet object including two replicates of
Hi-C contact counts.}

\item{dispF}{A dispersionFunction object.}
}
\value{
variance-stabilized InteractionSet object.
}
\description{
This function takes a learned dispersionFunction and stabilize the variance of
Hi-C contact counts from two replicates (provided as an
InteractionSet object)
}
